Add UserContactMethod unit tests and permission behaviour

+++ src/app/Services/UserContactMethodService.php
@@ -5,6 +5,7 @@ namespace App\Services;
 use App\Http\Requests\User\UserContactMethod\UserContactMethodStoreRequest;
 use App\Http\Requests\User\UserContactMethod\UserContactMethodUpdateRequest;
 use App\Models\UserContactMethod;
+use Illuminate\Support\Facades\Gate;
 
 class UserContactMethodService
 {
@@ -18,6 +19,8 @@ class UserContactMethodService
 
     public function store(UserContactMethodStoreRequest $request): UserContactMethod
     {
+        Gate::authorize('create', UserContactMethod::class);
+
         $validated = $request->validated();
 
         $userContactMethod = UserContactMethod::create([
@@ -32,13 +35,20 @@ class UserContactMethodService
 
     public function show(string $id): UserContactMethod
     {
-        return UserContactMethod::findOrFail($id);
+        $userContactMethod = UserContactMethod::find($id);
+
+        Gate::authorize('view', $userContactMethod);
+
+        return $userContactMethod;
     }
 
     public function update(UserContactMethodUpdateRequest $request, string $id): UserContactMethod
     {
+
         $validated = $request->validated();
         $userContactMethod = $this->show($id);
+
+        Gate::authorize('update', $userContactMethod);
 
         $userContactMethod->update($validated);
 
@@ -48,6 +58,8 @@ class UserContactMethodService
     public function delete(string $id): void
     {
         $userContactMethod = UserContactMethod::findOrFail($id);
+
+        Gate::authorize('delete', $userContactMethod);
 
         $userContactMethod->delete();
     }
